% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black.R
\name{style_active_file_black}
\alias{style_active_file_black}
\title{Style entire active file}
\usage{
style_active_file_black()
}
\description{
Assumes file is a standard RMarkdown or Quarto document with Python-labeled
code blocks.
}
\details{
Code credit: Alex Rossell Hayes (https://github.com/rossellhayes)
}
