% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black.R
\name{style_black}
\alias{style_black}
\title{Style Python Code directly}
\usage{
style_black(code)
}
\arguments{
\item{code}{character text of Python code}
}
\value{
character formatted Python code
}
\description{
Takes Python code directly into the function and formats it with black.
}
