% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black.R
\name{style_black_selection}
\alias{style_black_selection}
\title{Apply black formatting to highlighted code}
\usage{
style_black_selection()
}
\value{
None
}
\description{
Apply black formatting to highlighted code
}
