% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/black.R
\name{style_python_two_spaces}
\alias{style_python_two_spaces}
\title{Replace Python indentation from 4 to 2}
\usage{
style_python_two_spaces()
}
\value{
None
}
\description{
To give another option for consistency if you want to use 2 spaces for Python
}
